fix bug on line 41 -
even when error was thrown for the population sizes file it still entered the if clause because the value for input_population_sizes_file is not reset. I am not sure but fileInput might not be supported by shinyjs::reset(). so now the if clause is rather written with pop:sized data which is writable and I can set to NULL in case of error.